voice_to_text/requirements.txt
# Base dependencies
numpy<2
requests==2.32.3

# PyTorch CPU (corrected specification)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cpu

# Application specific
flask==3.0.2
gunicorn==21.2.0
pydub==0.25.1
openai-whisper==20231117
ffmpeg-python==0.2.0
deep-translator==1.11.4

# Required for whisper
regex==2024.11.6
tqdm==4.66.4